refactor: move state to homepage from cardlist
